requirements.txt
flask
flask-cors
openai
python-dotenv
openai==0.28
requests